prediction changing the precision of the results

+++ kre/reportregression.cpp
@@ -24,9 +24,10 @@ protected:
 	}
 };
 
-
-unsigned long long returnPred(vector<unsigned long long> v) {
-	unsigned long long sum = 0.0;
+//unsigned long long returnPred(vector<unsigned long long> V)
+// unsigned long long sum
+double returnPred(vector<double> v) {
+	double sum = 0.0;
 	for (int i = 0; i < v.size(); i++) {
 	    sum = sum + v.at(i);
 	}
@@ -147,14 +148,18 @@ void create_kre_report(string Ni, string bf, string fdist, string astar, string 
 	outputFile.open(output.c_str(), ios::out);
 
 	//We do know how to implement f-dist
-        vector<unsigned long long> vpred;
+        //vector<unsigned long long> vpred
+        vector<double> vpred;
  
 	for (int i = 0; i < v_astar_f.size(); i++) {
             int threshold = v_astar_f.at(i);
             //cout<<"threshold = "<<threshold<<endl;
 	    //outputFile<<threshold<<"\n\n";
-            unsigned long long pred = 0.0;
-            vector<unsigned long long> v_pred;
+ 
+            //unsigned long long pred = 0.0
+            //vector<unsigned long long> v_pred;
+            double pred = 0.0;
+            vector<double> v_pred;
             for (int j = 0; j < v_Ni.size(); j++) {
                 int g = j;
                 unsigned long long ni = v_Ni.at(j);
@@ -209,8 +214,8 @@ void create_kre_report(string Ni, string bf, string fdist, string astar, string 
 				} 
 				//outputFile<<"total nodes = "<<sumq<<"\n";
 				//outputFile<<"nodes with f less than or equal to"<<threshold<<" = "<<sumR<<"\n";
-
-                                long double percentage = 0;
+                                //long double percentage = 0;
+                                double percentage = 0;
                                 if (sumq > 0) {
                                    percentage =  (double)sumR/(double)sumq;
                                 } else {
@@ -219,7 +224,8 @@ void create_kre_report(string Ni, string bf, string fdist, string astar, string 
                                 cout<<"percentage = "<<percentage<<endl;       
 				//outputFile<<"percentage = "<<percentage<<"\n";
 				//cout<<"percentage = "<<percentage<<" ";
-			 	unsigned long long kre_i = ni*percentage;
+                                //unsigned long long kre_i = ni*percentage;
+			 	double kre_i = ni*percentage;
                                 v_pred.push_back(kre_i);
                                 m.clear();
 			} //end if
